+++ tests/data/tokenizers/sentence_splitter_test.py
@@ -1,3 +1,4 @@
+from allennlp.common import Params
 from allennlp.common.testing import AllenNlpTestCase
 from allennlp.data.tokenizers.sentence_splitter import SpacySentenceSplitter
 
@@ -45,3 +46,12 @@ class TestSentenceSplitter(AllenNlpTestCase):
             assert len(batch_doc) == len(separate_doc)
             for batch_sentence, separate_sentence in zip(batch_doc, separate_doc):
                 assert batch_sentence == separate_sentence
+
+    def test_to_params(self):
+        params = self.dep_parse_splitter.to_params()
+        assert isinstance(params, Params)
+        assert params.params == {
+            "type": "spacy",
+            "language": self.dep_parse_splitter._language,
+            "rule_based": self.dep_parse_splitter._rule_based,
+        }