+++ tests/data/tokenizers/spacy_tokenizer_test.py
@@ -1,5 +1,6 @@
 import spacy
 
+from allennlp.common import Params
 from allennlp.common.testing import AllenNlpTestCase
 from allennlp.data.tokenizers import Token, SpacyTokenizer
 
@@ -118,3 +119,19 @@ class TestSpacyTokenizer(AllenNlpTestCase):
         tokens = word_tokenizer.tokenize(sentence)
         assert tokens
         assert all(isinstance(token, spacy.tokens.Token) for token in tokens)
+
+    def test_to_params(self):
+        tokenizer = SpacyTokenizer()
+        params = tokenizer.to_params()
+        assert isinstance(params, Params)
+        assert params.params == {
+            "type": "spacy",
+            "language": tokenizer._language,
+            "pos_tags": tokenizer._pos_tags,
+            "parse": tokenizer._parse,
+            "ner": tokenizer._ner,
+            "keep_spacy_tokens": tokenizer._keep_spacy_tokens,
+            "split_on_spaces": tokenizer._split_on_spaces,
+            "start_tokens": tokenizer._start_tokens,
+            "end_tokens": tokenizer._end_tokens,
+        }